+++ .env.example
@@ -25,8 +25,11 @@ REDIS_URL=
 ENVIRONMENT=development
 LOG_LEVEL=INFO
 
+# CrewAI
+CREW_MEMORY_ENABLED=false
+
 # =============================================================================
-# Meta Ads CLI Integration
+# Meta Ads API Integration
 # =============================================================================
 # System user access token — generate in Meta Business Manager:
 #   Business Settings → System Users → <your system user> → Generate Token
@@ -64,4 +67,4 @@ GAM_JSON_KEY_PATH=
 GAM_APPLICATION_NAME=AAMPBuyerAgent
 
 # GAM API version (matches seller agent)
-GAM_API_VERSION=v202411
+GAM_API_VERSION=v202505
